Refactor and add entry tests
